airflow/requirements.txt
apache-airflow==2.9.3
boto3==1.34.0
psycopg2-binary==2.9.9
pandas==2.1.4
requests==2.31.0
apache-airflow-providers-docker==3.10.0
docker==6.1.3
urllib3==1.26.18
requests-unixsocket==0.3.0